requirements.txt
gspread
google-auth
pyyaml
yagmail[all]
flask
python-dotenv
schedule